+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0-alpha.0](https://github.com/coveo/ui-kit/compare/v0.5.1-alpha.1...v0.6.0-alpha.0) (2021-02-08)
+
+
+### Features
+
+* **atomic:** atomic-sort-dropdown completion ([#475](https://github.com/coveo/ui-kit/issues/475)) ([64ff6d8](https://github.com/coveo/ui-kit/commit/64ff6d8394b05bc03a50fe813a174ed40a8b7ff1))
+
+
+
+
+
+## [0.5.1-alpha.1](https://github.com/coveo/ui-kit/compare/v0.5.1-alpha.0...v0.5.1-alpha.1) (2021-02-05)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.5.0](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.7...v0.5.0) (2021-02-04)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
 # [0.5.0-alpha.6](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.5...v0.5.0-alpha.6) (2021-02-03)